Streamline psmomconfig.c by utilizing plugincommon

+++ plugins/psmom/src/psmom/psmom.c
@@ -193,16 +193,16 @@ static int shutdownJobs()
 */
 static void saveConfigValues()
 {
-    getConfParamI("TIME_OBIT", &obitTime);
-    getConfParamI("DEBUG_MASK", &debugMask);
+    obitTime = getConfValueI(&config, "TIME_OBIT");
+    debugMask = getConfValueI(&config, "DEBUG_MASK");
 
     /* network configuration */
-    getConfParamI("PORT_MOM", &momPort);
-    getConfParamI("PORT_RM", &rmPort);
-    getConfParamI("PORT_SERVER", &serverPort);
+    momPort = getConfValueI(&config, "PORT_MOM");
+    rmPort = getConfValueI(&config, "PORT_RM");
+    serverPort = getConfValueI(&config, "PORT_SERVER");
 
     /* torque protocol version */
-    getConfParamI("TORQUE_VERSION", &torqueVer);
+    torqueVer = getConfValueI(&config, "TORQUE_VERSION");
 }
 
 void startPsmom()
@@ -314,29 +314,28 @@ static bool initAccountingFunc()
  */
 static void cleanupLogs()
 {
-    int cleanJob, cleanNodes, cleanTemp;
+    bool cleanJob, cleanNodes, cleanTemp;
     char *dir = NULL;
 
     /* cleanup jobscript files */
-    getConfParamI("CLEAN_JOBS_FILES", &cleanJob);
+    cleanJob = getConfValueI(&config, "CLEAN_JOBS_FILES");
     if (cleanJob) {
-	dir = getConfParamC("DIR_JOB_FILES");
+	dir = getConfValueC(&config, "DIR_JOB_FILES");
 	removeDir(dir, 0);
     }
 
     /* cleanup node files */
-    getConfParamI("CLEAN_NODE_FILES", &cleanNodes);
+    cleanNodes = getConfValueI(&config, "CLEAN_NODE_FILES");
     if (cleanNodes) {
-	dir = getConfParamC("DIR_NODE_FILES");
+	dir = getConfValueC(&config, "DIR_NODE_FILES");
 	removeDir(dir, 0);
     }
 
     /* cleanup temp (scratch) dir */
-    getConfParamI("CLEAN_TEMP_DIR", &cleanTemp);
+    cleanTemp = getConfValueI(&config, "CLEAN_TEMP_DIR");
     if (cleanTemp) {
-	if ((dir = getConfParamC("DIR_TEMP"))) {
-	    removeDir(dir, 0);
-	}
+	dir = getConfValueC(&config, "DIR_TEMP");
+	if (dir) removeDir(dir, 0);
     }
 }
 
@@ -396,53 +395,55 @@ static int validateScripts()
     struct stat st;
     char *script, *dir, filename[400];
 
-    if ((script = getConfParamC("BACKUP_SCRIPT"))) {
-	if ((stat(script, &st)) == -1) {
+    script = getConfValueC(&config, "BACKUP_SCRIPT");
+    if (script) {
+	if (stat(script, &st) == -1) {
 	    mwarn(errno, "%s: invalid backup script '%s'", __func__, script);
 	    return 1;
 	}
 
-	if (((st.st_mode & S_IFREG) != S_IFREG) ||
-		((st.st_mode & S_IXUSR) != S_IXUSR)) {
+	if ((st.st_mode & S_IFREG) != S_IFREG
+	    || (st.st_mode & S_IXUSR) != S_IXUSR) {
 	    mlog("%s: invalid permissions for backup script '%s'\n",
-		    __func__, script);
+		 __func__, script);
 	    return 1;
 	}
     }
 
-    if ((script = getConfParamC("TIMEOUT_SCRIPT"))) {
-	if ((stat(script, &st)) == -1) {
+    script = getConfValueC(&config, "TIMEOUT_SCRIPT");
+    if (script) {
+	if (stat(script, &st) == -1) {
 	    mwarn(errno, "%s: invalid timeout script '%s'", __func__, script);
 	    return 1;
 	}
 
-	if (((st.st_mode & S_IFREG) != S_IFREG) ||
-		((st.st_mode & S_IXUSR) != S_IXUSR)) {
+	if ((st.st_mode & S_IFREG) != S_IFREG
+	    || (st.st_mode & S_IXUSR) != S_IXUSR) {
 	    mlog("%s: invalid permissions for timeout script '%s'\n",
-		    __func__, script);
+		 __func__, script);
 	    return 1;
 	}
     }
 
     /* validate prologue/epilogue scripts */
-    dir = getConfParamC("DIR_SCRIPTS");
+    dir = getConfValueC(&config, "DIR_SCRIPTS");
     snprintf(filename, sizeof(filename), "%s/prologue", dir);
-    if ((checkPELogueFileStats(filename, 1)) == -2) {
+    if (checkPELogueFileStats(filename, 1) == -2) {
 	mlog("%s: invalid permissions for '%s'\n", __func__, filename);
 	return 1;
     }
     snprintf(filename, sizeof(filename), "%s/prologue.parallel", dir);
-    if ((checkPELogueFileStats(filename, 1)) == -2) {
+    if (checkPELogueFileStats(filename, 1) == -2) {
 	mlog("%s: invalid permissions for '%s'\n", __func__, filename);
 	return 1;
     }
     snprintf(filename, sizeof(filename), "%s/epilogue", dir);
-    if ((checkPELogueFileStats(filename, 1)) == -2) {
+    if (checkPELogueFileStats(filename, 1) == -2) {
 	mlog("%s: invalid permissions for '%s'\n", __func__, filename);
 	return 1;
     }
     snprintf(filename, sizeof(filename), "%s/epilogue.parallel", dir);
-    if ((checkPELogueFileStats(filename, 1)) == -2) {
+    if (checkPELogueFileStats(filename, 1) == -2) {
 	mlog("%s: invalid permissions for '%s'\n", __func__, filename);
 	return 1;
     }
@@ -460,49 +461,48 @@ static int validateDirs()
     struct stat st;
     char *dir;
 
-    dir = getConfParamC("DIR_JOB_FILES");
-    if ((stat(dir, &st)) == -1 || ((st.st_mode & S_IFDIR) != S_IFDIR)) {
+    dir = getConfValueC(&config, "DIR_JOB_FILES");
+    if (stat(dir, &st) == -1 || (st.st_mode & S_IFDIR) != S_IFDIR) {
 	mwarn(errno, "%s: invalid job dir '%s'", __func__, dir);
 	return 1;
     }
 
-    dir = getConfParamC("DIR_NODE_FILES");
-    if ((stat(dir, &st)) == -1 || ((st.st_mode & S_IFDIR) != S_IFDIR)) {
+    dir = getConfValueC(&config, "DIR_NODE_FILES");
+    if (stat(dir, &st) == -1 || (st.st_mode & S_IFDIR) != S_IFDIR) {
 	mwarn(errno, "%s: invalid node files dir '%s'", __func__, dir);
 	return 1;
     }
 
-    if ((dir = getConfParamC("DIR_TEMP"))) {
-	if ((stat(dir, &st)) == -1 || ((st.st_mode & S_IFDIR) != S_IFDIR)) {
-	    mwarn(errno, "%s: invalid temp dir '%s'", __func__, dir);
-	    return 1;
-	}
+    dir = getConfValueC(&config, "DIR_TEMP");
+    if (dir && (stat(dir, &st) == -1 || (st.st_mode & S_IFDIR) != S_IFDIR)) {
+	mwarn(errno, "%s: invalid temp dir '%s'", __func__, dir);
+	return 1;
     }
 
-    dir = getConfParamC("DIR_JOB_ACCOUNT");
-    if ((stat(dir, &st)) == -1 || ((st.st_mode & S_IFDIR) != S_IFDIR)) {
+    dir = getConfValueC(&config, "DIR_JOB_ACCOUNT");
+    if (stat(dir, &st) == -1 || (st.st_mode & S_IFDIR) != S_IFDIR) {
 	mwarn(errno, "%s: invalid account dir '%s'", __func__, dir);
 	return 1;
     }
 
-    dir = getConfParamC("DIR_SCRIPTS");
-    if ((stat(dir, &st)) == -1 || ((st.st_mode & S_IFDIR) != S_IFDIR)) {
+    dir = getConfValueC(&config, "DIR_SCRIPTS");
+    if (stat(dir, &st) == -1 || (st.st_mode & S_IFDIR) != S_IFDIR) {
 	mwarn(errno, "%s: invalid scripts dir '%s'", __func__, dir);
 	return 1;
     }
 
     dir = DEFAULT_DIR_JOB_UNDELIVERED;
-    if ((stat(dir, &st)) == -1 || ((st.st_mode & S_IFDIR) != S_IFDIR)) {
+    if (stat(dir, &st) == -1 || (st.st_mode & S_IFDIR) != S_IFDIR) {
 	mwarn(errno, "%s: invalid undelivered dir '%s'", __func__, dir);
 	return 1;
     }
 
-    if ((dir = getConfParamC("DIR_LOCAL_BACKUP"))) {
-	if ((stat(dir, &st)) == -1 || ((st.st_mode & S_IFDIR) != S_IFDIR)) {
-	    mwarn(errno, "%s: invalid backup dir '%s'", __func__, dir);
-	    return 1;
-	}
+    dir = getConfValueC(&config, "DIR_LOCAL_BACKUP");
+    if (dir && (stat(dir, &st) == -1 || (st.st_mode & S_IFDIR) != S_IFDIR)) {
+	mwarn(errno, "%s: invalid backup dir '%s'", __func__, dir);
+	return 1;
     }
+
     return 0;
 }
 
@@ -745,7 +745,7 @@ void cleanup(void)
 
     if (!isInit) {
 	/* free config values */
-	clearConfig();
+	freeConfig(&config);
 	return;
     }
 
@@ -755,7 +755,7 @@ void cleanup(void)
     }
 
     /* free config values */
-    clearConfig();
+    freeConfig(&config);
 
     /* set collect mode in psaccount */
     if (psAccountSetGlobalCollect) psAccountSetGlobalCollect(0);
